recon page in 2 parts | cleaned few style file
