+++ web/src/pages/EditMission/CreateMission.tsx
@@ -21,7 +21,6 @@ import Feedbacks from '../../component/Feedback';
 import TopBar from '../../component/SideBar/TopBar';
 import SideBar from '../../component/SideBar/SideBar';
 import config from '../../config';
-import { Input } from './EditMission';
 
 export default function CreateMission() {
     const [Title, setTitle] = useState('');
