+++ web/src/pages/EditMission/Mission.scss
@@ -19,7 +19,7 @@
     font-family: Arial;
     font-style: "Poppins-regular";
     color: #737272;
-  }
+}
 
 .cancel-btn{
     background-color:transparent;
