+++ web/src/pages/MissionDetail/Recon.scss
@@ -1,0 +1,53 @@
+.mission-tool-line {
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    justify-content: flex-end;
+}
+
+.mission-borderBtn {
+    border: 1px var(--primary-btn) solid;
+    color: var(--primary-btn);
+    background-color: transparent;
+    cursor: pointer;
+    margin: 0;
+    margin-bottom: 1rem;
+}
+
+.recon_info {
+    width: 50%;
+    padding-top: 1rem;
+}
+
+.recon_container {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+}
+
+.wappa_input_container {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.wappa_input {
+    width: 80%;
+    height: 2.5rem;
+    border-radius: 4px;
+    border: 1px solid var(--light-gray);
+    font-size: 14px;
+    gap: 0.5rem;
+}
+
+.wappa_input:focus {
+    outline: none;
+    border-color: var(--gradient-primary-btn);
+    padding: 0 5px;
+}
+
+.wappa_search {
+    height: 2.5rem;
+    padding: auto;
+}
