+++ web/src/pages/Vulnerability/EditVuln.tsx
@@ -20,7 +20,6 @@ import TopBar from '../../component/SideBar/TopBar';
 import Feedbacks from '../../component/Feedback';
 import config from '../../config';
 import vulnerabilities from '../../assets/strings/en/vulnerabilities.json';
-import { Input } from '../EditMission/EditMission';
 
 const MAX_IMAGES = 4;
